test version

+++ app/src/main/java/ru/orehovai/weatherforboroda/model/weather/DayShort.java
@@ -28,7 +28,7 @@ public class DayShort {
     private Double windSpeed;
     @SerializedName("wind_gust")
     @Expose
-    private Integer windGust;
+    private Double windGust;
     @SerializedName("wind_dir")
     @Expose
     private String windDir;
@@ -113,11 +113,11 @@ public class DayShort {
         this.windSpeed = windSpeed;
     }
 
-    public Integer getWindGust() {
+    public Double getWindGust() {
         return windGust;
     }
 
-    public void setWindGust(Integer windGust) {
+    public void setWindGust(Double windGust) {
         this.windGust = windGust;
     }
