+++ app/src/main/java/ru/orehovai/weatherforboroda/model/weather/Night.java
@@ -37,7 +37,7 @@ public class Night {
     private Double windSpeed;
     @SerializedName("wind_gust")
     @Expose
-    private Integer windGust;
+    private Double windGust;
     @SerializedName("wind_dir")
     @Expose
     private String windDir;
@@ -149,11 +149,11 @@ public class Night {
         this.windSpeed = windSpeed;
     }
 
-    public Integer getWindGust() {
+    public Double getWindGust() {
         return windGust;
     }
 
-    public void setWindGust(Integer windGust) {
+    public void setWindGust(Double windGust) {
         this.windGust = windGust;
     }
